[Correção:Orders] Corrigido um erro que, ao reiniciar a aplicação as tabelas trocavam os IDs. Impossibilitando edição do arquivo salvo por falta de correspondência

+++ Wish.ERP.Benner/Views/OrderView.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
@@ -15,6 +16,7 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using Wish.ERP.Benner.Models;
 using Wish.ERP.Benner.Services;
+using Wish.ERP.Benner.ViewModels;
 using Wish.ERP.Benner.Views.Modals;
 using Wish.ERP.Benner.Views.Modals.Orders;
 
@@ -23,8 +25,6 @@ namespace Wish_ERP.Views
     public partial class OrderView : UserControl
     {
         private static OrderView instance = null;
-        private List<Order> selectedOrders;
-
 
         public static OrderView Instance
         {
@@ -40,34 +40,51 @@ namespace Wish_ERP.Views
         public OrderView()
         {
             DataContext = new Wish.ERP.Benner.ViewModels.OrdersViewModel();
+            
             InitializeComponent();
         }
         private void DeleteOrder(object sender, RoutedEventArgs e)
         {
-
-
-            var selectedOrders = OrdersList.SelectedItems.Cast<Order>().ToList();
-            if (selectedOrders.Count == 1)
+            if (DataContext is OrdersViewModel ovm)
             {
-                MessageBoxResult result = MessageBox.Show($"Tem certeza que deseja excluir {selectedOrders[0].Id}? Essa ação é Irreversível", "Atenção", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+                var selectedOrders = ovm.SelectedOrders;
 
-                if (result == MessageBoxResult.No) return;
-                OrdersServices.Delete(selectedOrders[0].Id);
-            }
-            else
-            {
-                MessageBoxResult result = MessageBox.Show($"Tem certeza que deseja excluir esses {selectedOrders.Count} clientes? Essa ação é Irreversível", "Atenção", MessageBoxButton.YesNo, MessageBoxImage.Warning);
-                if (result == MessageBoxResult.No) return;
-                var ids = selectedOrders.Select(c => c.Id).ToList();
-               OrdersServices.DeleteMany(ids);
+                if (!selectedOrders.Any()) return;
+
+                if (selectedOrders.Count == 1)
+                {
+                    var order = selectedOrders.First();
+                    MessageBoxResult result = MessageBox.Show(
+                        $"Tem certeza que deseja excluir {order.Id}? Essa ação é irreversível",
+                        "Atenção", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+
+                    if (result == MessageBoxResult.No) return;
+                    OrdersServices.Delete(order.Id);
+                }
+                else
+                {
+                    MessageBoxResult result = MessageBox.Show(
+                        $"Tem certeza que deseja excluir esses {selectedOrders.Count} clientes? Essa ação é irreversível",
+                        "Atenção", MessageBoxButton.YesNo, MessageBoxImage.Warning);
+
+                    if (result == MessageBoxResult.No) return;
+
+                    var ids = selectedOrders.Select(c => c.Id).ToList();
+                    OrdersServices.DeleteMany(ids);
+                }
             }
         }
+
 
         private void HandleSelectionList(object sender, SelectionChangedEventArgs e)
         {
-            selectedOrders = OrdersList.SelectedItems.Cast<Order>().ToList();
-            DeleteOrderButton.IsEnabled = selectedOrders.Any();
+            if (DataContext is OrdersViewModel ovm)
+            {
+                ovm.SelectedOrders = OrdersList.SelectedItems.Cast<Order>().ToList();
+                DeleteOrderButton.IsEnabled = ovm.SelectedOrders.Any();
+            }
         }
+
 
         private void CreateOrder(object sender, RoutedEventArgs e)
         {
@@ -81,6 +98,9 @@ namespace Wish_ERP.Views
             Regex regex = new Regex("[^0-9.]");
             e.Handled = regex.IsMatch(e.Text);
         }
+
+
+
     }
 }
